% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{eval_mapping_results}
\alias{eval_mapping_results}
\title{Evaluate mapping performance when using LOD drop or Bayesian credible support intervals}
\usage{
eval_mapping_results(
  thresh = seq(6, 9, by = 0.5),
  lod_drop = seq(1, 3, by = 0.1),
  bci = seq(0.8, 0.95, by = 0.01),
  use_lod_drop = TRUE,
  scans,
  map,
  true_qtl
)
}
\description{
This function ...
}
\examples{
eval_mapping_results()
}
