% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{eval_mapping_sampling_results}
\alias{eval_mapping_sampling_results}
\title{Evaluate mapping performance when using sampling procedures, such as bootstraps}
\usage{
eval_mapping_sampling_results(
  thresh = seq(6, 9, by = 0.5),
  prob = seq(0.8, 0.95, by = 0.01),
  parboot_results,
  num_true_qtl
)
}
\description{
This function ...
}
\examples{
eval_mapping_sampling_results()
}
