% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{eval_null_mapping_results}
\alias{eval_null_mapping_results}
\title{Evaluate mapping performance with null simulations}
\usage{
eval_null_mapping_results(thresh = seq(6, 10, by = 0.5), scans, map)
}
\description{
This function ...
}
\examples{
eval_null_mapping_results()
}
