% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_heritability_simulation_functions.R
\name{eval_sim_h2_miqtl_strainvar_add_only_fit}
\alias{eval_sim_h2_miqtl_strainvar_add_only_fit}
\title{Simulate data with additive and strain/genome heritability components and evaluate heritability estimation in an additive-only model}
\usage{
eval_sim_h2_miqtl_strainvar_add_only_fit(
  sim_h2_add_prop,
  h2_total,
  n_sims,
  n_per_strain,
  K_strains,
  K_strains_fit = NULL,
  intercept = 0,
  use_rint = FALSE,
  do_Kasv_fit = FALSE
)
}
\description{
This function ...
}
\examples{
eval_sim_h2_miqtl_strainvar_add_only_fit()
}
