% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_heritability_simulation_functions.R
\name{eval_sim_h2_sommer_strainvar}
\alias{eval_sim_h2_sommer_strainvar}
\title{Simulate data with additive and strain/genome heritability components and evaluate heritability estimation}
\usage{
eval_sim_h2_sommer_strainvar(
  sim_h2_add_prop,
  h2_total,
  n_sims,
  n_per_strain,
  K_strains,
  K_strains_fit = NULL,
  intercept = 0,
  use_rint = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{sim_h2_add_prop}{Specified proportion of heritability that is additive for the simulated data. Must be from 0 to 1.}

\item{h2_total}{Specified cumulative heritability for the simulated data. Must be from 0 to 1, though cannot be 1 exactly.}

\item{n_sims}{The number of simulations to perform.}

\item{n_per_strain}{The number of replicates per genome.}

\item{K_strains}{The strain-level genetic relationship (i.e., kinship) matrix used for simulation and by default for estimation.}

\item{K_strains_fit}{DEFAULT: NULL. Allows for a different kinship matrix to be provided for estimation.}

\item{intercept}{DEFAULT: 0. A shared intercept to center the data.}

\item{use_rint}{DEFAULT: FALSE. If TRUE, perfoms a rank-based inverse normal transformation on data after simulation but before estimation.}

\item{return_data}{DEFAULT: FALSE. IF TRUE, returns simulated data as well as heritability estimates.}
}
\value{
\code{eval_sim_h2_sommer_strainvar} returns a \code{n_sims}-by-\code{2} vector of heritability estimates if \code{return_data = FALSE}, otherwise returns a list with \code{n_sims}-vector and the \code{k}-by-\code{n_sims} simulated data matrix as elements.
}
\description{
This function simulates trait data for populations with replicates with a strain/genome heritability components and then performs heritability estimation.
}
\examples{
eval_sim_h2_sommer_strainvar(sim_h2_add_prop = 0.5, h2_total = 0.8, n_sims = 10, n_per_strain = 4, K_strains = K_CCstrain_overall)
}
