% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_plotting_functions.R
\name{make_ccrix_grid}
\alias{make_ccrix_grid}
\title{CC-RIX grid plot}
\usage{
make_ccrix_grid(
  cc_strains,
  ccrix_set,
  low = "white",
  high = "black",
  include_cc = TRUE,
  grid_line_width = 0.5,
  text_size = 6
)
}
\description{
This function ...
}
\examples{
make_ccrix_grid()
}
