% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_support_functions.R
\name{make_Kasv}
\alias{make_Kasv}
\title{Scale kinship matrix to average semivariance}
\usage{
make_Kasv(K = NULL, Z = NULL)
}
\arguments{
\item{K}{DEFAULT: NULL. Haplotype-based kinship matrix, as calculated with \code{qtl2::calc_kinship()}.}

\item{Z}{DEFAULT: NULL. SNP allele count design matrix. Provide Z if using a SNP-based kinship matrix.}
}
\description{
This function takes a kinship matrix and calculates the average semivariance form.
}
\examples{
make_Kasv()
}
