% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_plotting_functions.R
\name{plot_ccrix_grid}
\alias{plot_ccrix_grid}
\title{CC-RIX grid plot}
\usage{
plot_ccrix_grid(
  cc_strains,
  ccrix_set,
  low = "white",
  high = "black",
  include_cc = TRUE,
  grid_line_width = 0.5,
  text_size = 6
)
}
\arguments{
\item{cc_strains}{The set of CC strains that will make up the labels on each axes. Each element is expected to have the format "CCXXX".}

\item{ccrix_set}{The set of F1s from the stains in \code{cc_strains}. Each element is expected to have the format "CC001xCC002".}

\item{low}{DEFAULT: "white". The color of a cell that represents a missing F1.}

\item{high}{DEFAULT: "black". The color of a cell that represents an observed F1.}

\item{include_cc}{DEFAULT: TRUE. If TRUE, the CC strains (representing the diagonal) are colored as observed. If FALSE, they are colored as missing.}

\item{text_size}{DEFAULT: 6. The size of text for the labels.}

\item{grid_line_wide}{DEFAULT: 0.5. The line width of the grid lines.}
}
\description{
This function makes a grid plot to describe a CC-RIX sample
}
\examples{
cc_strains <- c(paste0("CC00", 1:9), paste0("CC0", 10:50))
cc_pairs <- combn(cc_strains, 2)
ccrix_names <- paste(cc_pairs[1,], cc_pairs[2,], sep = "x")
set.seed(10)
ccrix_set <- sample(ccrix_names, 100)
plot_ccrix_grid(cc_strains = cc_strains, ccrix_set = ccrix_set, include_cc = FALSE)
}
