% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_plotting_functions.R
\name{plot_founder_chromosomes}
\alias{plot_founder_chromosomes}
\title{Draw cartoon chromosome pair for the founder strains of the CC, CC-RIX, and DO}
\usage{
plot_founder_chromosomes(
  col,
  width = 0.15,
  gap = 0.3,
  y_shift = NULL,
  x_shift = 0,
  add_border = TRUE,
  border_lwd = 1.5,
  end_type = c("rounded", "circle", "square")
)
}
\arguments{
\item{col}{The color of the haplotype pair.}

\item{width}{DEFAULT: 0.15. The width of the chromatids.}

\item{gap}{DEFAULT: 0.3. The space between the chromatids.}

\item{y_shift}{DEFAULT: NULL. Option to shift the plot along the y-axis.}

\item{x_shift}{DEFAULT: 0. Option to shift the plot along the x-axis.}

\item{add_border}{DEFAULT: TRUE. If TRUE, a black outline is added to the chromatids. If FALSE, no outline is used.}

\item{border_lwd}{DEFAULT: 1.5. The line width of the outline.}

\item{end_type}{DEFAULT: "rounded". If "rounded", the corners of the chromatids are rounded. If "circle", the ends of the chromatids are half cirlces. If "square", the chromatids are rectangles with sharp corners.}
}
\description{
This function makes a founder haplotype pair cartoon (solid color).
}
\examples{
plot_founder_chromosomes(col = qtl2::CCcolors[1], width = 0.15, gap = 0.3)
}
