% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{pull_fwer_pval}
\alias{pull_fwer_pval}
\title{Pull FWER p-value from maximum LOD scores}
\usage{
pull_fwer_pval(maxlod, right_side = TRUE, lod_obs, use_gev = TRUE)
}
\arguments{
\item{maxlod}{Vector of the maximum LOD scores from each null bootstrap sample or permutation.}

\item{right_side}{DEFAULT: TRUE. If TRUE, p-value is calculated based on the right-hand tail of the distribution. If FALSE, left-hand tail is used.}

\item{lod_obs}{The observed peak LOD score from a QTL of interest.}

\item{use_gev}{DEFAULT: TRUE. If TRUE, maximum LOD scores are used to fit an extreme value distribution, which is used to calculate the threshold. If FALSE, empirical quantiles are used instead.}
}
\description{
This function estimates a genome-wide significant p-value, controlling family-wise error rates (FWER), based on maximum LOD scores from scans of null parametric bootstraps or permutations of the data and the LOD score of a QTL of interest.
}
\examples{
pull_fwer_pval()
}
