% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{run_parboot}
\alias{run_parboot}
\title{Run scans of parametric bootstrap samples to estimate a positional confidence interval for a QTL}
\usage{
run_parboot(
  qtl_table,
  sim_data,
  genoprobs,
  K,
  map,
  num_samples = 100,
  prob = seq(0.8, 0.99, by = 0.01),
  ignore_u = TRUE
)
}
\description{
This function ...
}
\examples{
run_parboot()
}
