% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{run_subsample}
\alias{run_subsample}
\title{Run scabs if subsamples of data to estimate a positional confidence interval for a QTL}
\usage{
run_subsample(
  qtl_table,
  sim_data,
  genoprobs,
  K,
  map,
  num_draws = 100,
  sample_perc = 2/3,
  prob = seq(0.8, 0.99, by = 0.01)
)
}
\description{
This function ...
}
\examples{
run_subsample()
}
