% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{eval_null_mapping_results}
\alias{eval_null_mapping_results}
\title{Evaluate mapping performance with null simulations}
\usage{
eval_null_mapping_results(thresh = seq(6, 10, by = 0.5), scans, map)
}
\arguments{
\item{thresh}{DEFAULT: seq(6, 9, by = 0.5). LOD score thresholds to be evaluated.}

\item{scans}{Simulated QTL scan output from \code{scan1}() of the qtl2 package.}

\item{map}{The marker/locus information corresponding to the genoprobs. The expected format is qtl2 (a per chromosome list of position vectors with marker/locus as element names).}
}
\description{
This function evaluates QTL performance in terms of false QTL counts for simulated null data.
}
\examples{
eval_null_mapping_results()
}
