% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_heritability_simulation_functions.R
\name{eval_sim_h2}
\alias{eval_sim_h2}
\title{Simulate data and evaluate heritability estimation}
\usage{
eval_sim_h2(
  sim_h2,
  n,
  K,
  K_fit = NULL,
  intercept = 0,
  method = c("qtl2", "sommer", "miqtl")
)
}
\description{
This function ...
}
\examples{
eval_sim_h2()
}
