% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_heritability_simulation_functions.R
\name{eval_sim_h2_sommer_strainvar}
\alias{eval_sim_h2_sommer_strainvar}
\title{Simulate data with additive and strain/genome heritability components and evaluate heritability estimation}
\usage{
eval_sim_h2_sommer_strainvar(
  sim_h2_add_prop,
  h2_total,
  n_sims,
  n_per_strain,
  K_strains,
  K_strains_fit = NULL,
  intercept = 0
)
}
\description{
This function ...
}
\examples{
eval_sim_h2_sommer_strainvar()
}
