% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_heritability_simulation_functions.R
\name{eval_sim_h2_with_reps}
\alias{eval_sim_h2_with_reps}
\title{Simulate data with strain/genome replicates and evaluate heritability estimation}
\usage{
eval_sim_h2_with_reps(
  sim_h2,
  n_sims,
  n_per_strain,
  K_strains,
  K_strains_fit = NULL,
  intercept = 0,
  method = c("sommer", "miqtl"),
  use_rint = FALSE,
  do_Kasv_sim = FALSE,
  do_Kasv_fit = FALSE
)
}
\description{
This function ...
}
\examples{
eval_sim_h2_with_reps()
}
