% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_heritability_simulation_functions.R
\name{eval_sim_h2_with_reps}
\alias{eval_sim_h2_with_reps}
\title{Simulate data with strain/genome replicates and evaluate heritability estimation}
\usage{
eval_sim_h2_with_reps(
  sim_h2,
  n_sims,
  n_per_strain,
  K_strains,
  K_strains_fit = NULL,
  intercept = 0,
  method = c("sommer", "miqtl"),
  use_rint = FALSE,
  do_Kasv_fit = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{sim_h2}{Specified heritability for the simulated data. Must be from 0 to 1, though cannot be 1 exactly.}

\item{n_sims}{The number of simulations to perform.}

\item{n_per_strain}{The number of replicates per genome.}

\item{K_strains}{The strain-level genetic relationship (i.e., kinship) matrix used for simulation and by default for estimation.}

\item{K_strains_fit}{DEFAULT: NULL. Allows for a different kinship matrix to be provided for estimation.}

\item{intercept}{DEFAULT: 0. A shared intercept to center the data.}

\item{method}{DEFAULT: "sommer". The statistical package used to estimate heritability. Options are "sommer" and "miqtl".}

\item{use_rint}{DEFAULT: FALSE. If TRUE, perfoms a rank-based inverse normal transformation on data after simulation but before estimation.}

\item{do_Kasv_fit}{DEFAULT: FALSE. If TRUE, calculates the ASV form of the individual-level kinship matrix.}

\item{return_data}{DEFAULT: FALSE. IF TRUE, returns simulated data as well as heritability estimates.}
}
\value{
\code{eval_sim_h2_with_reps} returns a \code{n_sims}-vector of heritability estimates if \code{return_data = FALSE}, otherwise returns a list with \code{n_sims}-vector and the \code{k}-by-\code{n_sims} simulated data matrix as elements.
}
\description{
This function simulates trait data for populations with replicates with a specified heritability and then performs heritability estimation.
}
\examples{
eval_sim_h2_with_reps(sim_h2 = 0.8, n_sims = 100, n_per_strain = 4, K_strains = K_CCstrain_overall, method = "miqtl")
}
