% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_plotting_functions.R
\name{plot_cc_chromosomes}
\alias{plot_cc_chromosomes}
\title{Draw cartoon chromosome pair for CC strains}
\usage{
plot_cc_chromosomes(
  intervals,
  cols,
  width = 0.1,
  gap = 0.05,
  y_shift = NULL,
  x_shift = 0,
  add_border = TRUE,
  border_lwd = 1.5,
  end_type = c("rounded", "circle", "square")
)
}
\description{
This function ...
}
\examples{
set.seed(10)
plot_cc_chromosomes(intervals = c(0, sort(runif(n = 15, min = 0, max = 1)), 1), 
col = sample(qtl2::CCcolors, 16, replace = TRUE), width = 0.15, gap = 0.3)
}
