% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_plotting_functions.R
\name{plot_do_chromosomes}
\alias{plot_do_chromosomes}
\title{Draw cartoon chromosome pair for DO or CC-RIX}
\usage{
plot_do_chromosomes(
  intervals1,
  intervals2,
  cols1,
  cols2,
  width = 0.1,
  gap = 0.05,
  y_shift = NULL,
  x_shift = 0,
  add_border = TRUE,
  border_lwd = 1.5,
  end_type = c("rounded", "circle", "square")
)
}
\description{
This function ...
}
\examples{
plot_do_chromosomes(c(0, sort(runif(15, 0, 1)), 1), c(0, sort(runif(15, 0, 1)), 1), cols1 = sample(qtl2::CCcolors, 16, replace = TRUE), cols2 = sample(qtl2::CCcolors, 16, replace = TRUE), width = 0.15, gap = 0.3) ## CC-RIX
plot_do_chromosomes(c(0, sort(runif(50, 0, 1)), 1), c(0, sort(runif(50, 0, 1)), 1), cols1 = sample(qtl2::CCcolors, 51, replace = TRUE), cols2 = sample(qtl2::CCcolors, 51, replace = TRUE), width = 0.15, gap = 0.3) ## DO
}
