% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_plotting_functions.R
\name{plot_founder_chromosomes}
\alias{plot_founder_chromosomes}
\title{Draw cartoon chromosome pair for the founder strains of the CC, CC-RIX, and DO}
\usage{
plot_founder_chromosomes(
  col,
  width = 0.1,
  gap = 0.05,
  y_shift = NULL,
  x_shift = 0,
  add_border = TRUE,
  border_lwd = 1.5,
  end_type = c("rounded", "circle", "square")
)
}
\description{
This function ...
}
\examples{
plot_founder_chromosomes(col = qtl2::CColors[1], width = 0.15, gap = 0.3)
}
