% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_plotting_functions.R
\name{plot_genome_scan}
\alias{plot_genome_scan}
\title{Genome scan plot}
\usage{
plot_genome_scan(
  qtl_scan,
  lodcolumn = 1,
  med_scan = NULL,
  med_highlight_dat = NULL,
  map,
  bgcolor = "white",
  altbgcolor = "white",
  bgcolor_alpha = 0.6,
  altbgcolor_alpha = 0.6,
  color = "cadetblue",
  altcolor = "cadetblue",
  medcol = "gray90",
  outlinecol = NA,
  hlines_col = "gray",
  hlines_lwd = 0.5,
  added_hline = NULL,
  annot_dat,
  target_symbol = NULL,
  chr = c(1:19, "X"),
  main = "",
  my_y_max = NULL,
  rug_col = "black"
)
}
\description{
This function ...
}
\examples{
plot_genome_scan()
}
