% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{pull_fwer_thresh}
\alias{pull_fwer_thresh}
\title{Pull FWER thresholds from maximum LOD scores}
\usage{
pull_fwer_thresh(maxlod, right_side = TRUE, fwer = 0.05, use_gev = TRUE)
}
\description{
This function ...
}
\examples{
pull_fwer_thresh()
}
