% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musppr_mapping_simulation_functions.R
\name{sim_mpp_data}
\alias{sim_mpp_data}
\title{Simulate MPP data with effects from a single QTL, strain, and kinship}
\usage{
sim_mpp_data(
  genoprobs,
  map,
  K = NULL,
  samples = NULL,
  qtl_effect_size = 0.1,
  strain_effect_size = 0,
  kinship_effect_size = 0.1,
  locus = NULL,
  vary_locus = TRUE,
  num_replicates = 1,
  num_sim = 1,
  M_ID = NULL,
  sample_method = c("uniform", "crp"),
  num_alleles = 8,
  scale_type = c("sample", "balanced"),
  num_founders = 8,
  beta = NULL,
  sim_label = "sim_y"
)
}
\description{
This function ...
}
\examples{
sim_mpp_data()
}
